Enabled all the upcoming features I'm aware of (except for `FullTypeThrows`, because it's currently badly broken in the Swift 6 compiler).
There were no real changes required, other than fixing an existing issue with a missing `any` on an existential (odd that the 5.10 compiler never complained about this, since it's been invalid Swift for several versions now… 🤔).

Unfortunately this requires bumping the minimum Swift version to 5.9, not because of any _actual_ code requirements for 5.9, but because the `enableExperimentalFeature` package Swift setting thingy is only available in 5.9 and later. 😔